ehr_migration_package/Stack_C/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dateutil==2.8.2
mysql-connector-python==8.2.0
PyJWT==2.8.0
APScheduler==3.10.1
requests==2.31.0
openai==1.3.0
google-generativeai==0.3.0
pdfplumber==0.10.0
python-docx==1.1.0
pypandoc==1.12
aiofiles==23.2.1








